bugfixes in HeatMapPO and HeatMapView
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@14840 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 408e9c66876457d0d7d16a60f482b947c9737c12

+++ application/DebsGrandChallenge2013.viewer/src/de/uniol/inf/is/odysseus/debs2013/viewer/HeatmapView.java
@@ -58,9 +58,9 @@ public class HeatmapView extends AbstractSoccerView implements IStreamEditorType
 			currentTuple = (Tuple<?>) element;
 			
 			int[] tempArray = {		(int)currentTuple.getAttribute(attributeIndexMap.get("cell_x1")),
-									((Double)currentTuple.getAttribute(attributeIndexMap.get("cell_y1"))).intValue(),
+									(int)currentTuple.getAttribute(attributeIndexMap.get("cell_y1")),
 									(int)currentTuple.getAttribute(attributeIndexMap.get("cell_x2")),
-									((Double)currentTuple.getAttribute(attributeIndexMap.get("cell_y2"))).intValue()};
+									(int)currentTuple.getAttribute(attributeIndexMap.get("cell_y2"))};
 			
 //			cellValues.put(tempArray, (Double)currentTuple.getAttribute(attributeIndexMap.get("percent_time_in_time_cell")));
